refactor(server): move insights codec to functions/insights.ts
The row<->domain seam gets no layer-noun. A layer-named directory builds a
shadow tree: for every entity there is a second file mirroring the entity
list without adding meaning. The identifiers already carry the contract --
decodeInsight/encodeInsightDefinition name the role, storedInsightDefinitionSchema
names the representation -- so the module is named for the entity and sits at
the canonical entity path beside dashboards.ts and drafts.ts.

This is symmetric with the framework's input side, where args is a field and
nobody names the layer. "Codec" stays as vocabulary in prose, never as a path.

Path-only: no identifiers change.
